[SYCL][libdevice] Move common group utils from sanitizer specific header
